+++ backend/src/main/java/com/apex/backend/dto/PaperSummaryDTO.java
@@ -2,13 +2,14 @@ package com.apex.backend.dto;
 
 import lombok.Builder;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 @Builder
 public class PaperSummaryDTO {
-    private double cash;
-    private double used;
-    private double free;
-    private double pnl;
+    private BigDecimal cash;
+    private BigDecimal used;
+    private BigDecimal free;
+    private BigDecimal pnl;
     private int positions;
 }
